IGNITE-17046 Move index classes to core
